requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
pydantic==2.9.2
numpy==2.1.3
pyyaml==6.0.2